+++ src/php/AppKernel.php
@@ -40,6 +40,7 @@ class AppKernel extends \Frontastic\Common\Kernel
             new \Frontastic\Common\ShopifyBundle\FrontasticCommonShopifyBundle(),
             new \Frontastic\Common\ShopwareBundle\FrontasticCommonShopwareBundle(),
             new \Frontastic\Common\SprykerBundle\FrontasticCommonSprykerBundle(),
+            new \Frontastic\Common\AlgoliaBundle\FrontasticCommonAlgoliaBundle(),
             new \Frontastic\Common\FindologicBundle\FrontasticCommonFindologicBundle(),
             new \Frontastic\Catwalk\FrontendBundle\FrontasticCatwalkFrontendBundle(),
             new \Frontastic\Catwalk\ApiCoreBundle\FrontasticCatwalkApiCoreBundle(),