+++ CHANGELOG.md
@@ -1,5 +1,16 @@
 # catwalk Changelog
 
+## `1.16.0` (2021-08-26)
+
+* chore: bumped version required of frontastic/common to 2.17.0
+* feat(FP-932): created scaffolding to perform Algolia search
+
+## `1.15.3` (2021-08-24)
+
+* fix: wrap response in array only when there is actions
+* fix: Custom Data Source removed environments, undeleted filter, cleanup on Gateway
+* fix: CustomDataSource and Facet refactor to use the UserBundle MetaData object
+
 ## `1.15.2` (2021-08-19)
 
 * fix: server build failure if @frontastic/theme-boost doesn't exist in node_modules